<fix>[sblk]: add timeout for lvm command with locking
add timeout for lvm command with locking to prevent the VG lock from being occupied due to other command blocking, like lvcreate/lvresize/lvremove..
If the lvm command with locking is hung, it will always occupy the lock and cannot be released.
And in scenarios where storage IO is slow and lock contention occurs, it may take longer to execute,
so we need to set a timeout that can tolerate this scenario.

Resolves: ZSV-8956
Resolves: ZSTAC-66089

Change-Id: I99e06f687468706d74787267616279626962679c