<fix>[kvmagent]: separately retrieve the value and id of PCI configurations
root cause:

lspci forces each field to be output separately
and imposes strict limits on field length.

If a field's content is too long
(e.g., the SDevice description of your Mellanox network card),
it will be truncated and represented with ..., causing the Subdevice ID to be hidden.

Separately retrieve the value and id of PCI configurations:

1 use lspci -Dmmv to obtain the value (descriptive information) of PCI devices.

2 use lspci -Dmmnv to obtain the id (numeric identifiers) of PCI devices.

Resolves: ZSTAC-76391

Change-Id: I6868776ceaedb412c35d4902808a57d83cfeefa1