<fix>[zstackctl]: Resolve the phantom lock issue
In dual-node mode, the VIP failover combined with an
uncaught exception during lock.unlock()
led to a phantom lock issue.

Resolves: ZSTAC-80498

Change-Id: I6f6d616c6a666d63776d65657a766d66636c7269

<fix>[zstackctl]: Resolve the phantom lock issue

In dual-node mode, the VIP failover combined with an
uncaught exception during lock.unlock()
led to a phantom lock issue.

Resolves: ZSTAC-80498

Change-Id: I6b74726a62726975746a666978786667667a7161

<fix>[zstackctl]: Resolve the phantom lock issue

In dual-node mode, the VIP failover combined with an
uncaught exception during lock.unlock()
led to a phantom lock issue.

Resolves: ZSTAC-80498

Change-Id: I6970706675757765786f796c6474746b79687a72